ocr_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
